+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2024-12-17
+### Fixed
+- `CI` integration tests for k8s.
+
+### Changed
+- Run unit tests for both **PromQL** and **MetricsQL**.
+
+### Added
+- `cmd/sloth/commands/k8scontroller` add **--victoriametrics** flag support.
+
 ## [0.13.1] - 2024-12-16
 ### Changed
 - Update dependencies versions a bit further.